aepp_sdk_elixir v0.1.0 AeppSDK.Contract

Contains all contract-related functionality.

In order for its functions to be used, a client must be defined first. Client example can be found at: AeppSDK.Client.new/4.

Types

call_options()

call_options() :: [
  fee: non_neg_integer(),
  ttl: non_neg_integer(),
  amount: non_neg_integer(),
  gas: non_neg_integer(),
  gas_price: non_neg_integer()
]

deploy_options()

deploy_options() :: [
  deposit: non_neg_integer(),
  amount: non_neg_integer(),
  gas: non_neg_integer(),
  gas_price: non_neg_integer(),
  fee: non_neg_integer(),
  ttl: non_neg_integer(),
  vm: :fate | :aevm
]

Functions

call(client, contract_address, source_code, function_name, function_args, opts \\ [])

call(
  AeppSDK.Client.t(),
  String.t(),
  String.t(),
  String.t(),
  [String.t()],
  call_options()
) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c(),
     return_value: String.t(),
     return_type: String.t(),
     log: [
       %{
         address: AeppSDK.Utils.Encoding.base58c(),
         data: String.t(),
         topics: [non_neg_integer()]
       }
     ]
   }}
  | {:error, String.t()}
  | {:error, Tesla.Env.t()}

Call a contract

Example

iex> contract_address = "ct_2sZ43ScybbzKkd4iFMuLJw7uQib1dpUB8VDi9pLkALV5BpXXNR"
iex> source_code = "contract Identity =
  datatype event = AddedNumberEvent(indexed int, string)

  record state = { number : int }

  entrypoint init(x : int) =
    { number = x }

  stateful entrypoint add_to_number(x : int) =
    Chain.event(AddedNumberEvent(x, "Added a number"))
    state.number + x"
iex> function_name = "add_to_number"
iex> function_args = ["33"]
iex> AeppSDK.Contract.call(client, contract_address, source_code, function_name, function_args)
{:ok,
 %{
   block_hash: "mh_2uzSrRdURXy4ATwCo3XpeSngH9ECXhkBj3MWEYFatqK4pJgFWG",
   block_height: 86362,
   log: [
     %{
       address: "ct_2XphkkmsJAbR4NbSpYFgHPgzfpveKRA9FFTJmp6jX8JRqnveeD",
       data: "Added a number",
       topics: [100006271334006235721916574864225776454052674644157840164656436983196903186403,
        33]
     }
   ],
   return_type: "ok",
   return_value: 75,
   tx_hash: "th_CpexcQGiM86HVtHR6HTzYc3HoakXW2Xjm77wVKctoZmxTH52u"
 }}

compile(source_code, vm \\ :fate)

Compile a contract

Example

iex> source_code = "contract Number =
  record state = { number : int }

  entrypoint init(x : int) =
    { number = x }

  entrypoint add_to_number(x : int) =
    state.number + x"
iex> AeppSDK.Contract.compile(source_code, :aevm)
{:ok,
 %{
   byte_code: <<98, 0, 0, 100, 98, 0, 0, 151, 145, 128, 128, 128, 81, 127, 112,
     194, 27, 63, 171, 248, 210, 119, 144, 238, 34, 30, 100, 222, 2, 111, 12,
     11, 11, 82, 86, 82, 53, 206, 145, 155, 60, 13, 206, 214, 183, 62, 20, 98,
     0, 0, 242, 87, 80, 128, 81, 127, 226, 35, 29, 108, 223, 201, 57, 22, 222,
     76, 179, 169, 133, 123, 246, 92, 244, 15, 194, 86, 244, 161, 73, 139, 63,
     126, 124, 152, 12, 25, 147, 68, 20, 98, 0, 0, 170, 87, 80, 96, 1, 25, 81,
     0, 91, 96, 0, 25, 89, 96, 32, 1, 144, 129, 82, 96, 32, 144, 3, 96, 0, 89,
     144, 129, 82, 129, 82, 89, 96, 32, 1, 144, 129, 82, 96, 32, 144, 3, 96, 3,
     129, 82, 144, 89, 96, 0, 81, 89, 82, 96, 0, 82, 96, 0, 243, 91, 96, 0, 128,
     82, 96, 0, 243, 91, 128, 96, 0, 81, 81, 1, 144, 80, 144, 86, 91, 96, 32, 1,
     81, 81, 131, 146, 80, 128, 145, 80, 80, 128, 89, 144, 129, 82, 89, 96, 32,
     1, 144, 129, 82, 96, 32, 144, 3, 96, 0, 25, 89, 96, 32, 1, 144, 129, 82,
     96, 32, 144, 3, 96, 0, 89, 144, 129, 82, 129, 82, 89, 96, 32, 1, 144, 129,
     82, 96, 32, 144, 3, 96, 3, 129, 82, 129, 82, 144, 80, 144, 86, 91, 96, 32,
     1, 81, 81, 144, 80, 89, 80, 128, 145, 80, 80, 98, 0, 0, 159, 86>>,
   compiler_version: "3.0.0",
   contract_source: 'contract Number =
     record state = { number : int }

     entrypoint init(x : int) =
       { number = x }

     entrypoint add_to_number(x : int) =
       state.number + x',
   type_info: [
     {<<112, 194, 27, 63, 171, 248, 210, 119, 144, 238, 34, 30, 100, 222, 2,
        111, 12, 11, 11, 82, 86, 82, 53, 206, 145, 155, 60, 13, 206, 214, 183,
        62>>, "add_to_number",
      <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 32, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 160, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>,
      <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 32, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>},
     {<<226, 35, 29, 108, 223, 201, 57, 22, 222, 76, 179, 169, 133, 123, 246,
        92, 244, 15, 194, 86, 244, 161, 73, 139, 63, 126, 124, 152, 12, 25, 147,
        68>>, "init",
      <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 32, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 160, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>,
      <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 32, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 160, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 192, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 1, 64, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 128, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0,
        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
        0, 0, 0, 0, 0, 0, 0>>}
   ]
 }}

create_calldata(source_code, function_name, function_args, vm \\ :fate)

Create contract calldata

Example

iex> source_code = "contract Number =
  record state = { number : int }

  entrypoint init(x : int) =
    { number = x }

  entrypoint add_to_number(x : int) =
    state.number + x"
iex> function_name = "init"
iex> function_args = ["42"]
iex> AeppSDK.Contract.create_calldata(source_code, function_name, function_args, :aevm)
{:ok,
   <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 32, 226, 35, 29, 108, 223, 201, 57, 22, 222, 76, 179, 169,
     133, 123, 246, 92, 244, 15, 194, 86, 244, 161, 73, 139, 63, 126, 124, 152,
     12, 25, 147, 68, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 42>>}

decode_return_value(return_value, return_type)

decode_return_value(String.t(), String.t()) :: {:ok, tuple()} | {:error, atom()}

Decode a return value

Example

iex> return_value = "cb_VNLOFXc="
iex> return_type = "ok"
iex> AeppSDK.Contract.decode_return_value(return_value, return_type)
{:ok, 42}

decode_return_value(sophia_type, return_value, return_type)

decode_return_value(String.t(), String.t(), String.t()) ::
  {:ok, tuple()} | {:error, atom()}

Decode a return value

Example

iex> sophia_type = "int"
iex> return_value = "cb_AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEvrXnzA"
iex> return_type = "ok"
iex> AeppSDK.Contract.decode_return_value(sophia_type, return_value, return_type)
{:ok, 75}

default_amount()

false

default_deposit()

false

default_gas()

false

default_gas_price()

false

deploy(client, source_code, init_args, opts \\ [])

deploy(AeppSDK.Client.t(), String.t(), [String.t()], deploy_options()) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c(),
     contract_id: AeppSDK.Utils.Encoding.base58c(),
     log: [
       %{
         address: AeppSDK.Utils.Encoding.base58c(),
         data: String.t(),
         topics: [non_neg_integer()]
       }
     ]
   }}
  | {:error, String.t()}
  | {:error, Tesla.Env.t()}

Deploy a contract

Example

iex> source_code = "contract Number =
  record state = { number : int }

  entrypoint init(x : int) =
    { number = x }

  entrypoint add_to_number(x : int) =
    state.number + x"
iex> init_args = ["42"]
iex> AeppSDK.Contract.deploy(client, source_code, init_args)
{:ok,
 %{
   block_hash: "mh_6fEZ9CCPNXxyjpKwSjkihv2UA5voRKCBpvrK24gK38zkZZB5Q",
   block_height: 86362,
   contract_id: "ct_2XphkkmsJAbR4NbSpYFgHPgzfpveKRA9FFTJmp6jX8JRqnveeD",
   log: [],
   return_type: "ok",
   return_value: "cb_Xfbg4g==",
   tx_hash: "th_CGCF321Sz8zWPMpSpa28gk3jDvvzeda8edhNhmnLgUvFYi14U"
 }}

encode_logs(logs, topic_types)

false

get(client, contract_address)

get(AeppSDK.Client.t(), String.t()) :: {:ok, map()} | {:error, String.t()}

Get contract information

Example

iex> contract_address = "ct_2SrKe33k4pHbXFuBp4q1dx2yEaiTQDyUoA2WqyAni1WuwjtZw"
iex> AeppSDK.Contract.get(client, contract_address)
{:ok,
 %{
   abi_version: 3,
   active: true,
   deposit: 0,
   id: "ct_2SrKe33k4pHbXFuBp4q1dx2yEaiTQDyUoA2WqyAni1WuwjtZw",
   owner_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
   referrer_ids: [],
   vm_version: 5
 }}

get_aci_function_return_type(source_code, function_name)

false

get_ct_version(opts)

false

get_function_return_type(source_code, function_name)

get_function_return_type(String.t(), String.t()) ::
  {:ok, String.t()} | {:error, String.t()}

Get the return type of a function in a contract

Example

iex> source_code = "contract Identity =
  record state = { number : int }

  entrypoint init(x : int) =
    { number = x }

  entrypoint add_to_number(x : int) =
    state.number + x"
iex> function_name = "add_to_number"
iex> AeppSDK.Contract.get_function_return_type(source_code, function_name)
{:ok, "int"}

get_vm(int)

false

typerep_decode_return_value(typerep, return_value, return_type)

false