aepp_sdk_elixir v0.5.3 AeppSDK.GeneralizedAccount

Contains all generalized accounts functionalities.

In order for its functions to be used, a client must be defined first. Client example can be found at: AeppSDK.Client.new/4.

For more information: https://github.com/aeternity/protocol/blob/master/generalized_accounts/generalized_accounts.md

Functions

attach(client, source_code, auth_fun, init_args, opts \\ [])

attach(AeppSDK.Client.t(), String.t(), String.t(), [String.t()], list()) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Attach a generalized account to a basic account. After a generalized account has been attached, it's possible pass an :auth option to transaction related functions in order to authorize them through the attached contract. A transaction is authorized whenever the call to the auth function returns true (and unauthorized when false).

The option looks like this: auth: [ auth_contract_source: "contract Authorization =

function auth(auth_value : bool) =
  auth_value",

auth_args: ["true"], fee: 1_000_000_000_000_00, gas: 50_000, gas_price: 1_000_000_000, ttl: 0 ] where gas, gas_price and ttl are optional.

Examples

iex> source_code = "contract Authorization =

  entrypoint auth(auth_value : bool) =
    auth_value"
iex> auth_fun = "auth"
iex> init_args = []
iex> AeppSDK.GeneralizedAccount.attach(client, source_code, auth_fun, init_args)
{:ok,
 %{
   block_hash: "mh_CfEuHm4V2omAQGNAxcdPARrkfnYbKuuF1HpGhG5oQvoVC34nD",
   block_height: 92967,
   tx_hash: "th_9LutrWD1FuFyx4MUUeMcfyF3uebfaP8t5gzatWDLyFYsqK744"
 }}

compute_auth_id(meta_tx)

compute_auth_id(map()) :: {:ok, binary()}

Computes an authorization id for given GA meta tx

Examples

iex> meta_tx = %{
                  abi_version: 3,
                  auth_data: <<43, 17, 244, 119, 202, 45, 27, 127>>,
                  fee: 100000000000000,
                  ga_id: "ak_wuLXPE5pd2rvFoxHxvenBgp459rW6Y1cZ6cYTZcAcLAevPE5M",
                  gas: 50000,
                  gas_price: 1000000,
                  ttl: 0,
                  tx: <<248, 87, 11, 1, 192, 184, 82, 248, 80, 12, 1, 161, 1, 124, 169, 154,
                    140, 216, 36, 178, 163, 239, 195, 198, 197, 213, 0, 88, 87, 19, 67, 5, 117,
                    212, 206, 105, 153, 178, 2, 203, 32, 248, 96, 25, 216, 161, 1, 11, 180, 237,
                    121, 39, 249, 123, 81, 225, 188, 181, 225, 52, 13, 18, 51, 91, 42, 43, 18,
                    200, 188, 82, 33, 214, 60, 75, 203, 57, 212, 30, 97, 100, 133, 3, 223, 210,
                    64, 0, 0, 0, 128>>
                }
iex> AeppSDK.GeneralizedAccount.compute_auth_id(meta_tx)
{:ok,
 <<141, 79, 64, 237, 32, 190, 35, 175, 230, 66, 224, 247, 43, 83, 109, 142, 1,
   161, 69, 1, 114, 107, 20, 99, 55, 155, 198, 212, 142, 147, 104, 117>>}

default_gas()

false