aepp_sdk_elixir v0.1.0 AeppSDK.Listener

A listener service that connects to peers in a network and notifies processes for new blocks and transactions.

Functions

check_tx_confirmations(client, tx_hash, block_count, subscriber_pid)

check_tx_confirmations(
  AeppSDK.Client.t(),
  AeppSDK.Utils.Encoding.base58c(),
  non_neg_integer(),
  pid()
) :: :ok | {:error, any()}

Subscribe a process to be notified whether or not a transaction with a given hash has been confirmed after a block_count amount of key blocks. Requires a client to be passed as the first argument as it makes requests to a node.

Example

iex> client = AeppSDK.Client.new(
  %{
    public: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
    secret:
      "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
  },
  "ae_uat",
  "https://sdk-testnet.aepps.com/v2",
  "https://sdk-testnet.aepps.com/v2"
)
iex> tx_hash = "th_2RMnW2GYbS2n5mwRtjFjb61ZgSNzNNUHGWU5EBmQXeUZfpP4vM"
iex> block_count = 10
iex> AeppSDK.Listener.check_tx_confirmations(client, tx_hash, block_count, self())
iex> flush()
{:tx_confirmations,
 %{
   height: 111283,
   status: :confirmed,
   tx_hash: "th_2RMnW2GYbS2n5mwRtjFjb61ZgSNzNNUHGWU5EBmQXeUZfpP4vM"
 }}
:ok

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

notify(event, _, _)

start(network_id, port \\ 3016)

start(String.t(), non_neg_integer()) :: {:ok, pid()}

Start the listener for mainnet or testnet (network_id - "ae_mainnet" or "ae_uat"). Starting the listener is required before subscribing to any events.

Example

iex> AeppSDK.Listener.start("ae_uat")
{:ok, #PID<0.261.0>}

start(peers, network_id, genesis_hash, port \\ 3016)

start(
  [String.t()],
  String.t(),
  AeppSDK.Utils.Encoding.base58c(),
  non_neg_integer()
) :: {:ok, pid()}

Start the listener for a custom/private network with a set of user defined peers, network ID and genesis hash. Starting the listener is required before subscribing to any events. A peer is defined in the following format: "aenode://peer_pubkey@host:port" i.e. "aenode://pp_2L8A5vSjnkLtfFNpJNgP9HbmGLD7ZAGFxoof47N8L4yyLAyyMi@18.136.37.63:3015"

Example

iex> AeppSDK.Listener.start(
      ["aenode://pp_2L8A5vSjnkLtfFNpJNgP9HbmGLD7ZAGFxoof47N8L4yyLAyyMi@18.136.37.63:3015"],
      "localnet",
      "kh_2Eo9AVWHxTKio278ccANnwtr9hkUpzb1nLePfUzs7StwWyJ2xB"
    )
{:ok, #PID<0.214.0>}

stop()

stop() :: :ok

Stop the listener

Example

iex> AeppSDK.Listener.stop()
:ok

subscribe(event, subscriber_pid)

subscribe(atom(), pid()) :: :ok

Subscribe a process to notifications for a general event. These events are the following (passed as an atom):

:micro_blocks

:key_blocks

:transactions

:pool_transactions

Micro blocks are in a light format, meaning that they don't contain any transactions. Subscribing to :transactions will notify for any mined transactions, while :pool_transactions will notify for any new transactions added to the pool.

Example

iex> AeppSDK.Listener.subscribe(:key_blocks, self())
iex> flush()
{:key_blocks,
  %{
   beneficiary: "ak_nv5B93FPzRHrGNmMdTDfGdd5xGZvep3MVSpJqzcQmMp59bBCv",
   height: 107613,
   info: "cb_AAAAAfy4hFE=",
   miner: "ak_V9dqj32R728EmdtaeXmGC7Zn94oAtje2mU2R9vQeCE3uDF52z",
   nonce: 15682880022608178725,
   pow_evidence: [36880953, 49948618, 71657014, 77943721, 103433465, 109309407,
    127422545, 140895507, 181678786, 182358531, 186372470, 188704699, 189292674,
    201708293, 215741852, 233410604, 255594855, 257736633, 266211649, 269325362,
    274670268, 292854435, 295392989, 307433512, 312822688, 330979144, 339358365,
    340662380, 356338119, 356413992, 363610991, 364811647, 387037152, 387943984,
    438683034, 469469067, 472181309, 479842054, 484418139, 506523084, 517149088,
    532705934],
   prev_hash: "mh_nxXxmga7AU5T3N7GGUV4s5T8YvJYEZnXGTaaUoLgBStva1dT4",
   prev_key_hash: "kh_j47UCW6UYYN4FhcMwQgynpAKdddKeAkPUUFyF6G35hArsfo8Q",
   root_hash: "bs_2hpHUMa7VWVTdcwwSm4L7U1CLMXqTbhrjZYJ2FMLz1LZCKfoCG",
   target: 504011610,
   time: 1562765121744,
   version: 3
 }}
:ok

subscribe(event, subscriber_pid, filter)

subscribe(atom(), pid(), AeppSDK.Utils.Encoding.base58c()) :: :ok

Subscribe a process to notifications for a filterable event i.e. a spend transaction with a specific recipient or an oracle query transaction that targets a specific oracle.

These events are the following and are filtered by a specific property:

:spend_transactions - recipient

:oracle_queries - oracle ID

:oracle_responses - oracle query ID

:pool_spend_transactions - recipient

:pool_oracle_queries - oracle ID

:pool_oracle_responses - oracle query ID

Example

iex> AeppSDK.Listener.subscribe(:spend_transactions, self(), "ak_2siRXKa5hT1YpR2oPwcU3LDpYzAdAcgt6HSNUt61NNV9NqkRP9")
iex> flush()
{:spend_transactions,
  %{
    hash: "th_2Z9198iTQbBBE3jdeFxJQNMHqGNSESxG7PyTs3vU3JZUjYs319",
    tx: %{
      amount: 100,
      fee: 16640000000,
      nonce: 3,
      payload: "",
      recipient_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
      sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
      ttl: 0,
      type: :spend_tx
    }
  }
}
:ok

subscribe_for_contract_events(client, subscriber_pid, contract_id)

subscribe_for_contract_events(
  AeppSDK.Client.t(),
  pid(),
  AeppSDK.Utils.Encoding.base58c()
) :: :ok

Subscribe a process to notifications for any events that are to be emitted by a contract.

Example

iex> client = AeppSDK.Client.new(
  %{
    public: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
    secret:
      "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
  },
  "ae_uat",
  "https://sdk-testnet.aepps.com/v2",
  "https://sdk-testnet.aepps.com/v2"
)
iex> contract_id = "ct_jzmNJ8ZrMRFw8boyhxFDVYZhNGkmBNNV8hhKeebnFoKFYoQ58"
iex> AeppSDK.Listener.subscribe_for_contract_events(client, self(), contract_id)
:ok

subscribe_for_contract_events(client, subscriber_pid, contract_id, event_name, topic_types \\ [])

subscribe_for_contract_events(
  AeppSDK.Client.t(),
  pid(),
  AeppSDK.Utils.Encoding.base58c(),
  String.t(),
  [atom()]
) :: :ok

Subscribe a process to notifications for any events with the specified name that are to be emitted by a contract. Optionally, topics can be encoded by passing a list of their respective types (must be in the same order as in the event). The types can be the following:

:address

:contract

:oracle

:oracle_query

:int

:bits

:bytes

:bool

Example

iex> client = AeppSDK.Client.new(
  %{
    public: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
    secret:
      "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
  },
  "ae_uat",
  "https://sdk-testnet.aepps.com/v2",
  "https://sdk-testnet.aepps.com/v2"
)
iex> contract_id = "ct_jzmNJ8ZrMRFw8boyhxFDVYZhNGkmBNNV8hhKeebnFoKFYoQ58"
iex> event_name = "SomeEvent"
iex> AeppSDK.Listener.subscribe_for_contract_events(client, self(), contract_id, event_name)
:ok

unsubscribe(event, subscriber_pid)

unsubscribe(atom(), pid()) :: :ok

unsubscribe(event, subscriber_pid, filter)

Unsubscribe a process from a filterable event.

Example

iex> AeppSDK.Listener.unsubscribe(:spend_transactions, self(), "ak_2siRXKa5hT1YpR2oPwcU3LDpYzAdAcgt6HSNUt61NNV9NqkRP9")
:ok