aepp_sdk_elixir v0.1.0 AeppSDK.Oracle

Module for oracle interaction, see: https://github.com/aeternity/protocol/blob/master/oracles/oracles.md.

In order for its functions to be used, a client must be defined first. Client example can be found at: AeppSDK.Client.new/4.

Types

oracle_options()

oracle_options() :: [fee: non_neg_integer(), ttl: non_neg_integer()]

query_options()

query_options() :: [
  fee: non_neg_integer(),
  ttl: non_neg_integer(),
  query_fee: non_neg_integer()
]

sophia_data()

sophia_data() :: any()

sophia_type()

sophia_type() :: String.t()

See https://github.com/aeternity/protocol/blob/master/contracts/sophia.md#types

ttl()

ttl() :: %{type: ttl_type(), value: non_neg_integer()}

ttl_type()

ttl_type() :: :relative | :absolute

Functions

extend(client, oracle_id, oracle_ttl, opts \\ [])

extend(
  AeppSDK.Client.t(),
  AeppSDK.Utils.Encoding.base58c(),
  non_neg_integer(),
  oracle_options()
) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Extend the TTL of an oracle by a relative amount.

Example

iex> oracle_id = "ok_4K1dYTkXcLwoUEat9fMgVp3RrG3HTD51v4VzszYDgt2MqxzKM"
iex> ttl = 10
iex> AeppSDK.Oracle.extend(client, oracle_id, ttl)
{:ok,
 %{
   block_hash: "mh_21HxnSLJRhqB9S3aUfLDAqR3BMFKPj62vT1zuy1MsS7N4Ps94s",
   block_height: 77276,
   tx_hash: "th_3911tboNbJWA6X57tejX8yGQALdeqAQECk1BwyS43pPtEXt4C"
 }}

get_oracle(client, oracle_id)

get_oracle(AeppSDK.Client.t(), AeppSDK.Utils.Encoding.base58c()) ::
  {:ok,
   %{
     id: AeppSDK.Utils.Encoding.base58c(),
     query_format: binary(),
     response_format: binary(),
     query_fee: non_neg_integer(),
     ttl: non_neg_integer(),
     abi_version: 0 | 1
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Get an oracle object.

Example

iex> oracle_id = "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> AeppSDK.Oracle.get_oracle(client, oracle_id)
{:ok,
 %{
   abi_version: 1,
   id: "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
   query_fee: 30,
   query_format: "cb_AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAATWiO1g==",
   response_format: "cb_AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAKAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAATWiO1g==",
   ttl: 77316
 }}

get_queries(client, oracle_id)

get_queries(AeppSDK.Client.t(), AeppSDK.Utils.Encoding.base58c()) ::
  {:ok,
   [
     %{
       fee: non_neg_integer(),
       id: AeppSDK.Utils.Encoding.base58c(),
       oracle_id: AeppSDK.Utils.Encoding.base58c(),
       query: any(),
       response: any(),
       response_ttl: non_neg_integer(),
       sender_id: AeppSDK.Utils.Encoding.base58c(),
       sender_nonce: non_neg_integer(),
       ttl: non_neg_integer()
     }
   ]}
  | {:error, String.t()}
  | {:error, Env.t()}

Get all queries of an oracle.

Example

iex> oracle_id = "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> AeppSDK.Oracle.get_queries(client, oracle_id)
{:ok,
 [
   %{
     fee: 30,
     id: "oq_253A8BSZqUofetC5U9DqdJfYAcF5SHi6DPr5gPPSrayP8cwSUP",
     oracle_id: "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
     query: %{"a" => 1},
     query_ttl: 83952,
     response: "",
     response_ttl: 1,
     sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
     sender_nonce: 11662
   }
 ]}

get_query(client, oracle_id, query_id)

get_query(
  AeppSDK.Client.t(),
  AeppSDK.Utils.Encoding.base58c(),
  AeppSDK.Utils.Encoding.base58c()
) ::
  {:ok,
   %{
     fee: non_neg_integer(),
     id: AeppSDK.Utils.Encoding.base58c(),
     oracle_id: AeppSDK.Utils.Encoding.base58c(),
     query: any(),
     response: any(),
     response_ttl: non_neg_integer(),
     sender_id: AeppSDK.Utils.Encoding.base58c(),
     sender_nonce: non_neg_integer(),
     ttl: non_neg_integer()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Get an oracle query by its ID.

Example

iex> oracle_id = "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> query_id = "oq_253A8BSZqUofetC5U9DqdJfYAcF5SHi6DPr5gPPSrayP8cwSUP"
iex> AeppSDK.Oracle.get_query(client, oracle_id, query_id)
{:ok,
 %{
   fee: 30,
   id: "oq_253A8BSZqUofetC5U9DqdJfYAcF5SHi6DPr5gPPSrayP8cwSUP",
   oracle_id: "ok_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
   query: %{"a" => 1},
   query_ttl: 83952,
   response: "",
   response_ttl: 1,
   sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
   sender_nonce: 11662
 }}

query(client, oracle_id, query, query_ttl, response_ttl_value, opts \\ [])

query(
  AeppSDK.Client.t(),
  AeppSDK.Utils.Encoding.base58c(),
  String.t(),
  ttl(),
  non_neg_integer(),
  query_options()
) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c(),
     query_id: AeppSDK.Utils.Encoding.base58c()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Query an oracle. Keep in mind that the response TTL is always relative, and that the sum of the relative query and response TTL can't be higher than the oracle's TTL.

Example

iex> oracle_id = "ok_4K1dYTkXcLwoUEat9fMgVp3RrG3HTD51v4VzszYDgt2MqxzKM"
iex> query = "a query"
iex> query_ttl = %{type: :relative, value: 1_000}
iex> response_ttl = 1_000
iex> AeppSDK.Oracle.query(client, oracle_id, query, query_ttl, response_ttl)
{:ok,
 %{
   block_hash: "mh_2SgSB1yeekq8JseSfkKhAPuvs7RF3YUjm896g4aj2GPpSa9AnJ",
   block_height: 77276,
   query_id: "oq_u7sgmMQNjZQ4ffsN9sSmEhzqsag1iEfx8SkHDeG1y8EbDB5Aq",
   tx_hash: "th_2esUdavCBmW1oYSCichdQv3txyWXYsDSAum2jAvQfpgktJ4oEt"
 }}

register(client, query_format, response_format, ttl, query_fee, opts \\ [])

register(
  AeppSDK.Client.t(),
  sophia_type(),
  sophia_type(),
  ttl(),
  non_neg_integer(),
  oracle_options()
) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c(),
     oracle_id: AeppSDK.Utils.Encoding.base58c()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Register a typed oracle. Queries and responses that don't follow the oracle's respective formats are invalid. The query and response types are sophia types.

Example

iex> query_format = "string"
iex> response_format = "map(string, string)"
iex> oracle_ttl = %{type: :relative, value: 10_000}
iex> query_fee = 100
iex> AeppSDK.Oracle.register(client, query_format, response_format, oracle_ttl, query_fee)
{:ok,
  %{
    block_hash: "mh_5zfVXCDwsBRjukTPjKRaS7T3TCc4Mn5PMTS19cWbcjRjeXjcF",
    block_height: 77276,
    oracle_id: "ok_4K1dYTkXcLwoUEat9fMgVp3RrG3HTD51v4VzszYDgt2MqxzKM",
    tx_hash: "th_21qrcDco5fL1cuaNqM1Ug1ojHiSzjnuEYzVEpwxVwuS2V95qBk"
  }}

respond(client, oracle_id, query_id, response, response_ttl, opts \\ [])

respond(
  AeppSDK.Client.t(),
  AeppSDK.Utils.Encoding.base58c(),
  AeppSDK.Utils.Encoding.base58c(),
  sophia_data(),
  non_neg_integer(),
  oracle_options()
) ::
  {:ok,
   %{
     block_hash: AeppSDK.Utils.Encoding.base58c(),
     block_height: non_neg_integer(),
     tx_hash: AeppSDK.Utils.Encoding.base58c()
   }}
  | {:error, String.t()}
  | {:error, Env.t()}

Respond to an oracle query. Only the oracle's owner can respond to its queries.

Example

iex> oracle_id = "ok_4K1dYTkXcLwoUEat9fMgVp3RrG3HTD51v4VzszYDgt2MqxzKM"
iex> query_id = "oq_u7sgmMQNjZQ4ffsN9sSmEhzqsag1iEfx8SkHDeG1y8EbDB5Aq"
iex> response = %{"a" => "response"}
iex> query_ttl = %{type: :relative, value: 1_000}
iex> response_ttl = 1_000
iex> AeppSDK.Oracle.respond(client, oracle_id, query_id, response, response_ttl)
{:ok,
 %{
   block_hash: "mh_QTXMDn8Ln6fiBBXByXJkEeD6wq6QzQZHMVuApbouTFaqWMkSt",
   block_height: 77276,
   tx_hash: "th_2cBk9pBEwMSD3xtYMvrsFWJ3atuGN29XQUZfKV65KYMQuPJLiV"
 }}