aepp_sdk_elixir v0.5.3 AeppSDK.Utils.Account

Account AeppSDK.Utils.

In order for its functions to be used, a client must be defined first. Client example can be found at: AeppSDK.Client.new/4.

Functions

next_valid_nonce(client, public_key)

next_valid_nonce(AeppSDK.Client.t(), String.t()) ::
  {:ok, integer()} | {:error, String.t()} | {:error, Tesla.Env.t()}

Get the next valid nonce for a public key

Example

iex> public_key = "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> AeppSDK.Utils.Account.next_valid_nonce(client, public_key)
{:ok, 8544}

nonce_at_hash(client, public_key, block_hash)

nonce_at_hash(AeppSDK.Client.t(), String.t(), String.t()) ::
  {:ok, integer()} | {:error, String.t()} | {:error, Tesla.Env.t()}

Get the nonce after a block indicated by hash

Example

iex> public_key = "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> block_hash = "kh_WPQzXtyDiwvUs54N1L88YsLPn51PERHF76bqcMhpT5vnrAEAT"
iex> AeppSDK.Utils.Account.nonce_at_hash(client, public_key, block_hash)
{:ok, 8327}