aepp_sdk_elixir v0.1.0 AeppSDK.Utils.Serialization

Transaction serialization.

Types

id()

id() :: {:id, id_type(), binary()}

id_type()

id_type() :: :account | :oracle | :name | :commitment | :contract | :channel

poi_keyword()

poi_keyword() :: [
  accounts: {state_hash(), map()},
  calls: {state_hash(), map()},
  channels: {state_hash(), map()},
  contracts: {state_hash(), map()},
  ns: {state_hash(), map()},
  oracles: {state_hash(), map()}
]

rlp_binary()

rlp_binary() :: binary()

state_hash()

state_hash() :: binary()

structure_type()

structure_type() ::
  :signed_tx
  | :spend_tx
  | :oracle_register_tx
  | :oracle_query_tx
  | :oracle_response_tx
  | :oracle_extend_tx
  | :name_claim_tx
  | :name_preclaim_tx
  | :name_update_tx
  | :name_revoke_tx
  | :name_transfer_tx
  | :contract_create_tx
  | :contract_call_tx
  | :sophia_byte_code
  | :channel_create_tx
  | :channel_deposit_tx
  | :channel_withdraw_tx
  | :channel_close_mutual_tx
  | :channel_close_solo_tx
  | :channel_slash_tx
  | :channel_settle_tx
  | :channel_snapshot_solo_tx
  | :channel_force_progress_tx
  | :transfer_update
  | :deposit_update
  | :withdraw_update
  | :create_contract_update
  | :call_contract_update

tx_type()

tx_type() ::
  AeternityNode.Model.SpendTx.t()
  | AeternityNode.Model.OracleRegisterTx.t()
  | AeternityNode.Model.OracleRespondTx.t()
  | AeternityNode.Model.OracleQueryTx.t()
  | AeternityNode.Model.OracleExtendTx.t()
  | AeternityNode.Model.NamePreclaimTx.t()
  | AeternityNode.Model.NameClaimTx.t()
  | AeternityNode.Model.NameRevokeTx.t()
  | AeternityNode.Model.NameTransferTx.t()
  | AeternityNode.Model.NameUpdateTx.t()
  | AeternityNode.Model.ContractCreateTx.t()
  | AeternityNode.Model.ContractCallTx.t()
  | AeternityNode.Model.ChannelCreateTx.t()
  | AeternityNode.Model.ChannelCloseMutualTx.t()
  | AeternityNode.Model.ChannelCloseSoloTx.t()
  | AeternityNode.Model.ChannelDepositTx.t()
  | AeternityNode.Model.ChannelForceProgressTx.t()
  | AeternityNode.Model.ChannelSettleTx.t()
  | AeternityNode.Model.ChannelSlashTx.t()
  | AeternityNode.Model.ChannelSnapshotSoloTx.t()
  | AeternityNode.Model.ChannelWithdrawTx.t()

Functions

deserialize(payload, type)

deserialize(binary(), structure_type()) :: list()

Deserialize an RLP binary payload with the template corresponding to the given type

Example

iex> payload = <<248, 224, 43, 1, 161, 1, 11, 180, 237, 121, 39, 249, 123, 81, 225, 188, 181,
    225, 52, 13, 18, 51, 91, 42, 43, 18, 200, 188, 82, 33, 214, 60, 75, 203, 57,
    212, 30, 97, 130, 33, 131, 161, 5, 64, 216, 143, 81, 41, 52, 245, 89, 135,
    253, 7, 12, 94, 142, 96, 251, 212, 96, 76, 248, 1, 152, 97, 16, 144, 62, 43,
    186, 148, 174, 76, 114, 1, 136, 27, 193, 109, 103, 78, 200, 0, 0, 0, 0, 131,
    15, 66, 64, 132, 59, 154, 202, 0, 184, 128, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 32, 112, 194, 27,
    63, 171, 248, 210, 119, 144, 238, 34, 30, 100, 222, 2, 111, 12, 11, 11, 82,
    86, 82, 53, 206, 145, 155, 60, 13, 206, 214, 183, 62, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 33>>
iex> type = :contract_call_tx
iex> AeppSDK.Utils.Serialization.deserialize(payload, type)
[
  caller_id: {:id, :account,
   <<11, 180, 237, 121, 39, 249, 123, 81, 225, 188, 181, 225, 52, 13, 18, 51,
     91, 42, 43, 18, 200, 188, 82, 33, 214, 60, 75, 203, 57, 212, 30, 97>>},
  nonce: 8579,
  contract_id: {:id, :contract,
   <<64, 216, 143, 81, 41, 52, 245, 89, 135, 253, 7, 12, 94, 142, 96, 251, 212,
     96, 76, 248, 1, 152, 97, 16, 144, 62, 43, 186, 148, 174, 76, 114>>},
  abi_version: 1,
  fee: 2000000000000000000,
  ttl: 0,
  amount: 0,
  gas: 1000000,
  gas_price: 1000000000,
  call_data: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 32, 112, 194, 27, 63, 171, 248, 210, 119, 144,
    238, 34, 30, 100, 222, 2, 111, 12, 11, 11, 82, 86, 82, 53, 206, 145, 155,
    60, 13, 206, 214, 183, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 33>>
]

id_to_record(value, type)

id_to_record(binary(), id_type()) :: id()

serialize(tx)

serialize(tx_type()) :: rlp_binary()

Serializes a transaction to a binary.

Example

iex> alias AeternityNode.Model.{Ttl,OracleRegisterTx}
iex> AeppSDK.Utils.Serialization.serialize( %OracleRegisterTx{
     query_format: <<"query_format">>,
     response_format: <<"response_format">>,
     query_fee: 10,
     oracle_ttl: %Ttl{type: :absolute, value: 10},
     account_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
     nonce: 37122,
     fee: 0,
     ttl: 10,
     abi_version: 0x30001
   })
 <<248, 77, 22, 1, 161, 1, 9, 51, 126, 98, 138, 255, 218, 224, 184, 180, 31, 234,
  251, 255, 59, 141, 224, 214, 250, 79, 248, 30, 246, 237, 55, 83, 153, 134,
  240, 138, 216, 129, 130, 145, 2, 140, 113, 117, 101, 114, 121, 95, 102, 111,
  ...>>

serialize(fields, type)

serialize(list(), structure_type()) :: rlp_binary()

Serializes a list of fields with the template corresponding to the given type

Example

iex> fields = [{:id, :account,
    <<11, 180, 237, 121, 39, 249, 123, 81, 225, 188, 181, 225, 52, 13, 18, 51,
      91, 42, 43, 18, 200, 188, 82, 33, 214, 60, 75, 203, 57, 212, 30, 97>>},
   8579,
   {:id, :contract,
    <<64, 216, 143, 81, 41, 52, 245, 89, 135, 253, 7, 12, 94, 142, 96, 251, 212,
      96, 76, 248, 1, 152, 97, 16, 144, 62, 43, 186, 148, 174, 76, 114>>},
   1,
   2_000_000_000_000_000_000,
   0,
   0,
   1_000_000,
   1_000_000_000,
   <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 32, 112, 194, 27, 63, 171, 248, 210, 119, 144, 238, 34,
     30, 100, 222, 2, 111, 12, 11, 11, 82, 86, 82, 53, 206, 145, 155, 60, 13,
     206, 214, 183, 62, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 33>>]
    iex> type = :contract_call_tx
    iex> AeppSDK.Utils.Serialization.serialize(fields, type)
    <<248, 224, 43, 1, 161, 1, 11, 180, 237, 121, 39, 249, 123, 81, 225, 188, 181,
    225, 52, 13, 18, 51, 91, 42, 43, 18, 200, 188, 82, 33, 214, 60, 75, 203, 57,
    212, 30, 97, 130, 33, 131, 161, 5, 64, 216, 143, 81, 41, 52, 245, 89, 135,
    253, 7, 12, 94, 142, 96, 251, 212, 96, 76, 248, 1, 152, 97, 16, 144, 62, 43,
    186, 148, 174, 76, 114, 1, 136, 27, 193, 109, 103, 78, 200, 0, 0, 0, 0, 131,
    15, 66, 64, 132, 59, 154, 202, 0, 184, 128, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 32, 112, 194, 27,
    63, 171, 248, 210, 119, 144, 238, 34, 30, 100, 222, 2, 111, 12, 11, 11, 82,
    86, 82, 53, 206, 145, 155, 60, 13, 206, 214, 183, 62, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 96, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 33>>

serialize_for_client(tx, type)

serialize_poi(state_hashes_list)

serialize_poi(poi_keyword()) :: binary()

Serializes a Proof-of-Inclusion to a binary.

Example

iex> poi = [
    accounts: {<<148, 112, 174, 4, 233, 253, 29, 65, 31, 17, 40, 153, 167, 18, 37,
    179, 35, 141, 97, 95, 132, 162, 158, 78, 103, 11, 230, 39, 208, 27, 176,111>>, %{cache: {0, nil}}}
   ]
iex> AeppSDK.Utils.Serialization.serialize_poi poi
<<235, 60, 1, 227, 226, 160, 148, 112, 174, 4, 233, 253, 29, 65, 31,
17, 40, 153, 167, 18, 37, 179, 35, 141, 97, 95, 132, 162, 158, 78,
103, 11, 230, 39, 208, 27, 176, 111, 192, 192, 192, 192, 192, 192>>