aepp_sdk_elixir v0.5.3 AeppSDK.Utils.Transaction

Transaction AeppSDK.Utils.

In order for its functions to be used, a client must be defined first. Client example can be found at: AeppSDK.Client.new/4.

Types

tx_types()

tx_types() ::
  AeternityNode.Model.SpendTx.t()
  | AeternityNode.Model.OracleRegisterTx.t()
  | AeternityNode.Model.OracleQueryTx.t()
  | AeternityNode.Model.OracleRespondTx.t()
  | AeternityNode.Model.OracleExtendTx.t()
  | AeternityNode.Model.NamePreclaimTx.t()
  | AeternityNode.Model.NameClaimTx.t()
  | AeternityNode.Model.NameTransferTx.t()
  | AeternityNode.Model.NameRevokeTx.t()
  | AeternityNode.Model.NameUpdateTx.t()
  | AeternityNode.Model.ContractCallTx.t()
  | AeternityNode.Model.ContractCreateTx.t()

Functions

calculate_fee(tx, height, network_id, fee, gas_price)

calculate_fee(
  tx_types(),
  non_neg_integer(),
  String.t(),
  atom() | non_neg_integer(),
  non_neg_integer()
) :: non_neg_integer()

Calculate the fee of the transaction.

Example

iex> spend_tx = %AeternityNode.Model.SpendTx{
                    amount: 40_000_000,
                    fee: 0,
                    nonce: 10_624,
                    payload: "",
                    recipient_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
                    sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
                    ttl: 0
                  }
iex> AeppSDK.Utils.Transaction.calculate_fee(spend_tx, 51_900, "ae_uat", 0, 1_000_000)
16660000000

calculate_min_fee(tx, height, network_id)

calculate_min_fee(struct(), non_neg_integer(), String.t()) ::
  non_neg_integer() | {:error, String.t()}

Calculates minimum fee of given transaction, depends on height and network_id

Example

iex> name_pre_claim_tx = %AeternityNode.Model.NamePreclaimTx{
                             account_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
                             commitment_id: "cm_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
                             fee: 0,
                             nonce: 0,
                             ttl: 0
                           }
iex> AeppSDK.Utils.Transaction.calculate_min_fee(name_pre_claim_tx, 50_000, "ae_mainnet")
16500000000

calculate_n_times_fee(tx, height, network_id, fee, gas_price, times)

calculate_n_times_fee(
  tx_types(),
  non_neg_integer(),
  String.t(),
  non_neg_integer(),
  non_neg_integer(),
  non_neg_integer()
) :: non_neg_integer()

Calculates fee for given transaction n times.

Example

iex> spend_tx = %AeternityNode.Model.SpendTx{
                    amount: 1_000_000_000_000,
                    fee: 0,
                    nonce: 1,
                    payload: "",
                    recipient_id: "ak_wuLXPE5pd2rvFoxHxvenBgp459rW6Y1cZ6cYTZcAcLAevPE5M",
                    sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
                    ttl: 0
                  }
iex> AeppSDK.Utils.Transaction.calculate_n_times_fee(spend_tx, 58_336, "ae_uat", 0, 1_000_000, 5)
16740000000

default_await_attempt_interval()

default_await_attempt_interval() :: non_neg_integer()

default_await_attempts()

default_await_attempts() :: non_neg_integer()

default_fee_calculation_times()

default_fee_calculation_times() :: non_neg_integer()

default_payload()

default_payload() :: String.t()

default_ttl()

default_ttl() :: non_neg_integer()

dummy_fee()

dummy_fee() :: non_neg_integer()

gas_limit(tx, height)

gas_limit(struct(), non_neg_integer()) ::
  non_neg_integer() | {:error, String.t()}

Returns gas limit for given transaction, depends on height.

Example

iex> oracle_register_tx = %AeternityNode.Model.OracleRegisterTx{
                              abi_version: 0x60001,
                              account_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
                              fee: 0,
                              nonce: 37_122,
                              oracle_ttl: %AeternityNode.Model.Ttl{type: :absolute, value: 10},
                              query_fee: 10,
                              query_format: "query_format",
                              response_format: "response_format",
                              ttl: 10
                            }
iex> AeppSDK.Utils.Transaction.gas_limit(oracle_register_tx, 5)
16581

min_gas(tx, height)

min_gas(struct(), non_neg_integer()) :: non_neg_integer() | {:error, String.t()}

Calculates minimum gas needed for given transaction, also depends on height.

Example

iex> spend_tx = %AeternityNode.Model.SpendTx{
                    amount: 5_018_857_520_000_000_000,
                    fee: 0,
                    nonce: 37181,
                    payload: "",
                    recipient_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
                    sender_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
                    ttl: 0
                  }
iex> AeppSDK.Utils.Transaction.min_gas(spend_tx, 50_000)
16740

post(client, tx, auth_opts, signatures_list)

post(AeppSDK.Client.t(), tx_types(), atom() | list(), list() | atom()) ::
  {:ok, map()} | {:error, String.t()} | {:error, Tesla.Env.t()}

Serialize the list of fields to RLP transaction binary, sign it with the private key and network ID, add calculated minimum fee and post it to the node.

Example

iex> spend_tx = %AeternityNode.Model.SpendTx{
                    amount: 10_000,
                    fee: 16_680_000_000,
                    nonce: 1,
                    payload: "",
                    recipient_id: "ak_wuLXPE5pd2rvFoxHxvenBgp459rW6Y1cZ6cYTZcAcLAevPE5M",
                    sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
                    ttl: 0
                  }
iex> AeppSDK.Utils.Transaction.post(client, spend_tx, :no_auth, :one_signature)
{:ok,
 %{
   block_hash: "mh_2wRRkfzcHd24cGbqdqaLAhxgpv4iMB8y1Cp5n9FAfhvDZJ7Qh",
   block_height: 149,
   tx_hash: "th_umEMGk2S1EtkeAZCVHXDoTqQMdMawK9R9j1yvDZWjvKmstg5c"
 }}

sign_tx(tx, client, auth_opts \\ :no_opts)

sign_tx(tx_types(), AeppSDK.Client.t(), list() | :no_opts) ::
  {:ok, list()} | {:error, String.t()}

Signs the transaction.

Example

iex> spend_tx = %AeternityNode.Model.SpendTx{
                    amount: 1_000_000_000_000,
                    fee: 0,
                    nonce: 1,
                    payload: "",
                    recipient_id: "ak_wuLXPE5pd2rvFoxHxvenBgp459rW6Y1cZ6cYTZcAcLAevPE5M",
                    sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
                    ttl: 0
                  }
iex> AeppSDK.Utils.Transaction.sign_tx(spend_tx, client)
{:ok,
 [
   %AeternityNode.Model.SpendTx{
     amount: 1000000000000,
     fee: 0,
     nonce: 1,
     payload: "",
     recipient_id: "ak_wuLXPE5pd2rvFoxHxvenBgp459rW6Y1cZ6cYTZcAcLAevPE5M",
     sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
     ttl: 0
   },
   <168, 173, 62, 192, 198, 163, 29, 224, 83, 235, 68, 152, 82, 151, 77, 68,
    228, 138, 200, 45, 15, 74, 125, 222, 45, 150, 114, 79, 84, 160, 83, 255,
    63, 165, 45, 146, 164, 168, 144, 116, 200, 253, 80, 119, 6, 147, ...>>
 ]}