vdev AeternityNode.Api.Account

API calls for all endpoints tagged Account.

Functions

get_account_by_pubkey(connection, pubkey, opts \\ [])

get_account_by_pubkey(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_account_by_pubkey_and_hash(connection, pubkey, hash, opts \\ [])

get_account_by_pubkey_and_hash(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key after the block indicated by hash

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_account_by_pubkey_and_height(connection, pubkey, height, opts \\ [])

get_account_by_pubkey_and_height(
  Tesla.Env.client(),
  String.t(),
  integer(),
  keyword()
) :: {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key after the opening key block of the generation at height

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

height (integer()): The height

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_pending_account_transactions_by_pubkey(connection, pubkey, opts \\ [])

get_pending_account_transactions_by_pubkey(
  Tesla.Env.client(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.GenericTxs.t()} | {:error, Tesla.Env.t()}

Get pending account transactions by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.GenericTxs{}} on success {:error, info} on failure