aepp_sdk_elixir v0.1.0 AeternityNode.Api.External

API calls for all endpoints tagged External.

Functions

get_account_by_pubkey(connection, pubkey, opts \\ [])

get_account_by_pubkey(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_account_by_pubkey_and_hash(connection, pubkey, hash, opts \\ [])

get_account_by_pubkey_and_hash(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key after the block indicated by hash

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_account_by_pubkey_and_height(connection, pubkey, height, opts \\ [])

get_account_by_pubkey_and_height(
  Tesla.Env.client(),
  String.t(),
  integer(),
  keyword()
) :: {:ok, AeternityNode.Model.Account.t()} | {:error, Tesla.Env.t()}

Get an account by public key after the opening key block of the generation at height

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

height (integer()): The height

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Account{}} on success {:error, info} on failure

get_channel_by_pubkey(connection, pubkey, opts \\ [])

get_channel_by_pubkey(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.Channel.t()} | {:error, Tesla.Env.t()}

Get channel by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The pubkey of the channel

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Channel{}} on success {:error, info} on failure

get_contract(connection, pubkey, opts \\ [])

get_contract(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.ContractObject.t()} | {:error, Tesla.Env.t()}

Get a contract by pubkey

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The pubkey of the contract

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.ContractObject{}} on success {:error, info} on failure

get_contract_code(connection, pubkey, opts \\ [])

get_contract_code(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.ByteCode.t()} | {:error, Tesla.Env.t()}

Get contract code by pubkey

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The pubkey of the contract

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.ByteCode{}} on success {:error, info} on failure

get_contract_po_i(connection, pubkey, opts \\ [])

get_contract_po_i(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.PoI.t()} | {:error, Tesla.Env.t()}

Get a proof of inclusion for a contract

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): Contract pubkey to get proof for

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.PoI{}} on success {:error, info} on failure

get_contract_store(connection, pubkey, opts \\ [])

get_contract_store(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.ContractStore.t()} | {:error, Tesla.Env.t()}

Get contract store by pubkey

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The pubkey of the contract

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.ContractStore{}} on success {:error, info} on failure

get_current_generation(connection, opts \\ [])

get_current_generation(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.Generation.t()} | {:error, Tesla.Env.t()}

Get the current generation

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Generation{}} on success {:error, info} on failure

get_current_key_block(connection, opts \\ [])

get_current_key_block(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.KeyBlock.t()} | {:error, Tesla.Env.t()}

Get the current key block

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.KeyBlock{}} on success {:error, info} on failure

get_current_key_block_hash(connection, opts \\ [])

get_current_key_block_hash(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.InlineResponse200.t()} | {:error, Tesla.Env.t()}

Get the hash of the current key block

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.InlineResponse200{}} on success {:error, info} on failure

get_current_key_block_height(connection, opts \\ [])

get_current_key_block_height(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.InlineResponse2001.t()} | {:error, Tesla.Env.t()}

Get the height of the current key block

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.InlineResponse2001{}} on success {:error, info} on failure

get_generation_by_hash(connection, hash, opts \\ [])

get_generation_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.Generation.t()} | {:error, Tesla.Env.t()}

Get a generation by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the generation

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Generation{}} on success {:error, info} on failure

get_generation_by_height(connection, height, opts \\ [])

get_generation_by_height(Tesla.Env.client(), integer(), keyword()) ::
  {:ok, AeternityNode.Model.Generation.t()} | {:error, Tesla.Env.t()}

Get a generation by height

Parameters

connection (AeternityNode.Connection): Connection to server

height (integer()): The height

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Generation{}} on success {:error, info} on failure

get_key_block_by_hash(connection, hash, opts \\ [])

get_key_block_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.KeyBlock.t()} | {:error, Tesla.Env.t()}

Get a key block by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.KeyBlock{}} on success {:error, info} on failure

get_key_block_by_height(connection, height, opts \\ [])

get_key_block_by_height(Tesla.Env.client(), integer(), keyword()) ::
  {:ok, AeternityNode.Model.KeyBlock.t()} | {:error, Tesla.Env.t()}

Get a key block by height

Parameters

connection (AeternityNode.Connection): Connection to server

height (integer()): The height

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.KeyBlock{}} on success {:error, info} on failure

get_micro_block_header_by_hash(connection, hash, opts \\ [])

get_micro_block_header_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.MicroBlockHeader.t()} | {:error, Tesla.Env.t()}

Get a micro block header by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.MicroBlockHeader{}} on success {:error, info} on failure

get_micro_block_transaction_by_hash_and_index(connection, hash, index, opts \\ [])

get_micro_block_transaction_by_hash_and_index(
  Tesla.Env.client(),
  String.t(),
  integer(),
  keyword()
) :: {:ok, AeternityNode.Model.GenericSignedTx.t()} | {:error, Tesla.Env.t()}

Get a micro block transaction by hash and index

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the block

index (integer()): The index of the transaction in a block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.GenericSignedTx{}} on success {:error, info} on failure

get_micro_block_transactions_by_hash(connection, hash, opts \\ [])

get_micro_block_transactions_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.GenericTxs.t()} | {:error, Tesla.Env.t()}

Get micro block transactions by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.GenericTxs{}} on success {:error, info} on failure

get_micro_block_transactions_count_by_hash(connection, hash, opts \\ [])

get_micro_block_transactions_count_by_hash(
  Tesla.Env.client(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.InlineResponse2002.t()} | {:error, Tesla.Env.t()}

Get micro block transaction count by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the block

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.InlineResponse2002{}} on success {:error, info} on failure

get_name_entry_by_name(connection, name, opts \\ [])

get_name_entry_by_name(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.NameEntry.t()} | {:error, Tesla.Env.t()}

Get name entry from naming system

Parameters

connection (AeternityNode.Connection): Connection to server

name (String.t): The name key of the name entry

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.NameEntry{}} on success {:error, info} on failure

get_oracle_by_pubkey(connection, pubkey, opts \\ [])

get_oracle_by_pubkey(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.RegisteredOracle.t()} | {:error, Tesla.Env.t()}

Get an oracle by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the oracle

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.RegisteredOracle{}} on success {:error, info} on failure

get_oracle_queries_by_pubkey(connection, pubkey, opts \\ [])

get_oracle_queries_by_pubkey(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.OracleQueries.t()} | {:error, Tesla.Env.t()}

Get oracle queries by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the oracle

opts (KeywordList): [optional] Optional parameters

:from (String.t): Last query id in previous page

:limit (integer()): Max number of oracle queries

:type (String.t): The type of a query: open, closed or all

Returns

{:ok, %AeternityNode.Model.OracleQueries{}} on success {:error, info} on failure

get_oracle_query_by_pubkey_and_query_id(connection, pubkey, query_id, opts \\ [])

get_oracle_query_by_pubkey_and_query_id(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.OracleQuery.t()} | {:error, Tesla.Env.t()}

Get an oracle query by public key and query ID

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the oracle

query_id (String.t): The ID of the query

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.OracleQuery{}} on success {:error, info} on failure

get_peer_pubkey(connection, opts \\ [])

get_peer_pubkey(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.PeerPubKey.t()} | {:error, Tesla.Env.t()}

Get peer public key

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.PeerPubKey{}} on success {:error, info} on failure

get_pending_account_transactions_by_pubkey(connection, pubkey, opts \\ [])

get_pending_account_transactions_by_pubkey(
  Tesla.Env.client(),
  String.t(),
  keyword()
) :: {:ok, AeternityNode.Model.GenericTxs.t()} | {:error, Tesla.Env.t()}

Get pending account transactions by public key

Parameters

connection (AeternityNode.Connection): Connection to server

pubkey (String.t): The public key of the account

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.GenericTxs{}} on success {:error, info} on failure

get_pending_key_block(connection, opts \\ [])

get_pending_key_block(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.KeyBlock.t()} | {:error, Tesla.Env.t()}

Get the pending key block

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.KeyBlock{}} on success {:error, info} on failure

get_status(connection, opts \\ [])

get_status(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.Status.t()} | {:error, Tesla.Env.t()}

Get the status of a node

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.Status{}} on success {:error, info} on failure

get_top_block(connection, opts \\ [])

get_top_block(Tesla.Env.client(), keyword()) ::
  {:ok, AeternityNode.Model.KeyBlockOrMicroBlockHeader.t()}
  | {:error, Tesla.Env.t()}

Get the top block (either key or micro block)

Parameters

connection (AeternityNode.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.KeyBlockOrMicroBlockHeader{}} on success {:error, info} on failure

get_transaction_by_hash(connection, hash, opts \\ [])

get_transaction_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.GenericSignedTx.t()} | {:error, Tesla.Env.t()}

Get a transaction by hash

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the transaction

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.GenericSignedTx{}} on success {:error, info} on failure

get_transaction_info_by_hash(connection, hash, opts \\ [])

get_transaction_info_by_hash(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, AeternityNode.Model.TxInfoObject.t()} | {:error, Tesla.Env.t()}

Parameters

connection (AeternityNode.Connection): Connection to server

hash (String.t): The hash of the transaction

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.TxInfoObject{}} on success {:error, info} on failure

post_transaction(connection, body, opts \\ [])

post_transaction(Tesla.Env.client(), AeternityNode.Model.Tx.t(), keyword()) ::
  {:ok, AeternityNode.Model.PostTxResponse.t()} | {:error, Tesla.Env.t()}

Post a new transaction

Parameters

connection (AeternityNode.Connection): Connection to server

body (Tx): The new transaction

opts (KeywordList): [optional] Optional parameters

Returns

{:ok, %AeternityNode.Model.PostTxResponse{}} on success {:error, info} on failure