aepp_sdk_elixir v0.1.0 Core.Client

Contains the Client structure, holding all the data that is needed in order to use the SDK.

Types

keypair()

keypair() :: %{public: String.t(), secret: String.t()}

t()

t() :: %Core.Client{
  connection: struct(),
  gas_price: non_neg_integer(),
  internal_connection: struct(),
  keypair: keypair(),
  network_id: String.t()
}

Functions

new(keypair, network_id, url, internal_url, gas_price \\ 1000000)

new(keypair(), String.t(), String.t(), String.t(), non_neg_integer()) ::
  Core.Client.t()

Client constructor

Example

iex> public_key = "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> secret_key = "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
iex> network_id = "ae_uat"
iex> url = "https://sdk-testnet.aepps.com/v2"
iex> internal_url = "https://sdk-testnet.aepps.com/v2"
iex> Core.Client.new(%{public: public_key, secret: secret_key}, network_id, url, internal_url)
%Core.Client{
  connection: %Tesla.Client{
    adapter: nil,
    fun: nil,
    post: [],
    pre: [
      {Tesla.Middleware.BaseUrl, :call, ["https://sdk-testnet.aepps.com/v2"]}
    ]
  },
  gas_price: 0,
  internal_connection: %Tesla.Client{
    adapter: nil,
    fun: nil,
    post: [],
    pre: [
      {Tesla.Middleware.BaseUrl, :call, ["https://sdk-testnet.aepps.com/v2"]}
    ]
  },
  keypair: %{
    secret: "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61",
    public: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
  },
  network_id: "ae_uat"
}