vdev Utils.Transaction

Transaction utils

Types

tx_types()

tx_types() ::
  AeternityNode.Model.SpendTx.t()
  | AeternityNode.Model.OracleRegisterTx.t()
  | AeternityNode.Model.OracleQueryTx.t()
  | AeternityNode.Model.OracleRespondTx.t()
  | AeternityNode.Model.OracleExtendTx.t()
  | AeternityNode.Model.NamePreclaimTx.t()
  | AeternityNode.Model.NameClaimTx.t()
  | AeternityNode.Model.NameTransferTx.t()
  | AeternityNode.Model.NameRevokeTx.t()
  | AeternityNode.Model.NameUpdateTx.t()
  | AeternityNode.Model.ContractCallTx.t()
  | AeternityNode.Model.ContractCreateTx.t()

Functions

calculate_fee(tx, height, network_id, fee, gas_price)

calculate_fee(
  tx_types(),
  non_neg_integer(),
  String.t(),
  atom() | non_neg_integer(),
  non_neg_integer()
) :: non_neg_integer()

Calculate the fee of the transaction.

Examples

iex> spend_tx = %AeternityNode.Model.SpendTx{
  amount: 40000000,
  fee: 0,
  nonce: 10624,
  payload: "",
  recipient_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
  sender_id: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
  ttl: 0
  }
iex> Utils.Transaction.calculate_fee(spend_tx, 51_900, "ae_uat", 0, 1_000_000)
16660000000

calculate_min_fee(tx, height, network_id)

calculate_min_fee(struct(), non_neg_integer(), String.t()) ::
  non_neg_integer() | {:error, String.t()}

Calculates minimum fee of given transaction, depends on height and network_id

Examples:

 iex> name_pre_claim_tx =
        %AeternityNode.Model.NamePreclaimTx{
        account_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
        commitment_id: "cm_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
        fee: 0,
        nonce: 0,
        ttl: 0
      }
 iex> Utils.Transaction.calculate_min_fee(name_pre_claim_tx, 50000, "ae_mainnet")
   16500000000

default_await_attempt_interval()

default_await_attempt_interval() :: non_neg_integer()

default_await_attempts()

default_await_attempts() :: non_neg_integer()

default_payload()

default_payload() :: String.t()

default_ttl()

default_ttl() :: non_neg_integer()

dummy_fee()

dummy_fee() :: non_neg_integer()

gas_limit(tx, height)

gas_limit(struct(), non_neg_integer()) ::
  non_neg_integer() | {:error, String.t()}

Returns gas limit for given transaction, depends on height.

Examples:

 iex> oracle_register_tx =
     %AeternityNode.Model.OracleRegisterTx{
       abi_version: 196609,
       account_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
       fee: 0,
       nonce: 37122,
       oracle_ttl: %AeternityNode.Model.Ttl{type: :absolute, value: 10},
       query_fee: 10,
       query_format: "query_format",
       response_format: "response_format",
       ttl: 10
     }
 iex> Utils.Transaction.gas_limit oracle_register_tx, 5
    16581

min_gas(tx, height)

min_gas(struct(), non_neg_integer()) :: non_neg_integer() | {:error, String.t()}

Calculates minimum gas needed for given transaction, also depends on height.

Examples:

 iex> spend_tx =
        %AeternityNode.Model.SpendTx{
         amount: 5018857520000000000,
         fee: 0,
         nonce: 37181,
         payload: "",
         recipient_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
         sender_id: "ak_542o93BKHiANzqNaFj6UurrJuDuxU61zCGr9LJCwtTUg34kWt",
         ttl: 0
        }
 iex> Utils.Transaction.min_gas spend_tx, 50000
   16740

try_post(connection, secret_key, network_id, gas_price, tx, height)

try_post(
  struct(),
  String.t(),
  String.t(),
  non_neg_integer(),
  tx_types(),
  non_neg_integer()
) :: {:ok, map()} | {:error, String.t()} | {:error, Tesla.Env.t()}

Serialize the list of fields to an RLP transaction binary, sign it with the private key and network ID, add calculated minimum fee and post it to the node

Examples

iex> connection = AeternityNode.Connection.new("https://sdk-testnet.aepps.com/v2")
iex> public_key = "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU"
iex> secret_key = "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
iex> network_id = "ae_uat"
iex> gas_price = 1_000_000_000_000
iex> {:ok, nonce} = Utils.Account.next_valid_nonce(connection, public_key)
iex> source_code = "contract Number =

record state = { number : int }

function init(x : int) =

{ number = x }

function add_to_number(x : int) = state.number + x"

iex> function_name = "init"
iex> function_args = ["42"]
iex> {:ok, calldata} = Core.Contract.create_calldata(source_code, function_name, function_args)
iex> source_hash_bytes = 32
iex> {:ok, source_hash} = :enacl.generichash(source_hash_bytes, source_code)
iex> {:ok, %{byte_code: byte_code, type_info: type_info}} = Core.Contract.compile(source_code)
iex> byte_code_fields = [
  source_hash,
  type_info,
  byte_code
]
iex> serialized_wrapped_code = Utils.Serialization.serialize(byte_code_fields, :sophia_byte_code)
iex> contract_create_tx = %AeternityNode.Model.ContractCreateTx{
  owner_id: public_key,
  nonce: nonce,
  code: serialized_wrapped_code,
  abi_version: :unused,
  deposit: 0,
  amount: 0,
  gas: 1_000_000,
  gas_price: 1_000_000_000,
  fee: 0,
  ttl: Utils.Transaction.default_ttl(),
  call_data: calldata
}
iex> {:ok, %{height: height}} = AeternityNode.Api.Chain.get_current_key_block_height(connection)
iex> Utils.Transaction.try_post(connection, secret_key, network_id, gas_price, contract_create_tx, height)
{:ok,
 %{
   block_hash: "mh_29ZNDHkaa1k54Gr9HqFDJ3ubDg7Wi6yJEsfuCy9qKQEjxeHdH4",
   block_height: 68240,
   hash: "th_gfVPUw5zerDAkokfanFrhoQk9WDJaCdbwS6dGxVQWZce7tU3j"
 }}